Import NZ Freshwater Fish datasets.

Import data from the NZ Freshwater Fish Database. Enter search terms as arguments as you would in the NZFFD and import directly into R. You can search using all the same query options which are used for the NZFFFD, see their help page for details. To import the entire database leave all arguments blank.

nzffd_import(
  catchment = "",
  river = "",
  location = "",
  fish_method = "",
  species = "",
  starts = 1850,
  ends = 2100
)

Arguments

catchment
catchment number. a 6 digit number unique to the reach of interest. You can search using the indidviual number (e.g. catchment = "702.500"), or for all rivers in a catchment you can use the wildcard search term (e.g. catchment = "702%"), or don't set the arg if you want all catchments in NZ.

river
river name. e.g. to get all records for the Clutha, river = "Clutha".

location
sampling locality. e.g. location = "Awakino". This only works when the location is included in the name of the waterway.

fish_method
fishing method used. There are 59 different possible options for fish_method, if you want to search for a specific fishing method look at the dataset ?nzffd_method to see a list of all possible options, you can then copy/paste from there (e.g. if we only wanted fish caught be lures use fish_meth = "Angling - Lure") don't set the arg if you want all fishing methods.

species
species of interest.There are 75 different possible options for species, use ?nzffd_species function to see a list of all possible options. You can search using either common or scientific names and can search for multiple species at once. e.g. to search for Black mudfish use species = "Black mudfish" or species = "Neochanna diversus" and to search for Black mudfish and Bluegill bully use species = c("Black mudfish", "Bluegill bully") etc.

starts
start year, 1850 at the earliest.

ends
end year.

Value

A dataframe consisting of 22 columns where each row is a record for an individual species. The number of rows will depend on search terms.

Details

This function requires an internet connection to query NIWA's database.

Data citation: Crow S (2017). New Zealand Freshwater Fish Database. Version 1.2. The National Institute of Water and Atmospheric Research (NIWA). Occurrence Dataset https://doi.org/10.15468/ms5iqu

Examples

if (FALSE) {
df <- nzffd_import(starts = 2000, ends = 2001)
}